+++ src/civilization/general/mod.rs
@@ -1,2 +1,3 @@
 pub mod plugin;
 pub mod systems;
+pub mod components;
