+++ src/lib.rs
@@ -23,7 +23,7 @@ use crate::civilization::general::plugin::CivilizationPlugin;
 // This example game uses States to separate logic
 // See https://bevy-cheatbook.github.io/programming/states.html
 // Or https://github.com/bevyengine/bevy/blob/main/examples/ecs/state.rs
-#[derive(States, Default, Clone, Eq, PartialEq, Debug, Hash)]
+#[derive(States, Clone, PartialEq, Eq, Hash, Debug, Default)]
 enum GameState {
     // During the loading State the LoadingPlugin will load our assets
     #[default]
